+++ src/framework/piral-core/src/types/api.ts
@@ -3,7 +3,7 @@ import type { RouteComponentProps } from 'react-router';
 import type { PiletApi, Pilet, PiletMetadata, EventEmitter, SinglePilet, MultiPilet } from 'piral-base';
 import type { PiletCustomApi, PiralCustomPageMeta } from './custom';
 import type { AnyComponent } from './components';
-import type { ExtensionSlotProps, PiralExtensionSlotMap } from './extension';
+import type { ExtensionParams, ExtensionSlotProps, PiralExtensionSlotMap } from './extension';
 import type { SharedData, DataStoreOptions } from './data';
 import type { Disposable } from './utils';
 
@@ -49,7 +49,12 @@ export interface RouteBaseProps<UrlParams = any, UrlState = any>
 /**
  * The props used by a page component.
  */
-export interface PageComponentProps<T = any, S = any> extends RouteBaseProps<T, S> {}
+export interface PageComponentProps<T = any, S = any> extends RouteBaseProps<T, S> {
+  /**
+   * The meta data registered with the page.
+   */
+  meta: PiralPageMeta;
+}
 
 /**
  * The meta data registered for a page.
@@ -99,7 +104,7 @@ export interface PiletCoreApi {
   registerExtension<TName>(
     name: TName extends string ? TName : string,
     Component: AnyComponent<ExtensionComponentProps<TName>>,
-    defaults?: TName,
+    defaults?: Partial<ExtensionParams<TName>>,
   ): RegistrationDisposer;
   /**
    * Unregisters a global extension component.