+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Improved support and documentation for wrappers
 - Improved default metadata using `piletConfig` from *package.json* in pilet debug (#462)
+- Updated vulnerable dependencies
 - Fixed typing of extension slot defaults
 - Added new prop `meta` to registered pages
 - Added support for *debug-meta.json* when using `pilet debug` (#462)